feat(dde): open prospect fiche in a modal, place calls through Ringover
"Ouvrir" in the call file now opens the prospect's full fiche in a
window instead of scrolling back to the top of the page: same actions
(Appeler, Statut, Prendre rendez-vous, comment) and the same legacy CRM
data, without losing the reader's place in the list.

"Appeler" now hands the number to Ringover rather than the machine's own
dialer. The dialer URL is a template so it can be corrected without a
code change (NEXT_PUBLIC_RINGOVER_CALL_URL, {numero} = the number in
E.164). The number is copied to the clipboard at the same time, so the
call can still be placed by hand if the dialer does not prefill.

Prospect ids now come back from Postgres as integers rather than
bigint strings, matching the declared type they are compared against.

+++ lib/dde-prospects-db.ts
@@ -6,8 +6,8 @@ import type { DdeSession } from "./dde";
 import { DDE_PROSPECT_STATUT_KEYS, type DdeProspect } from "./dde-prospects";
 
 const P_COLS = `
-  id, crm_id, nom, prenom, telephone, telephone_2, email, adresse, code_postal, ville, departement,
-  statut, telepro_email, telepro_name, notes, appels, dernier_appel_at, rdv_id,
+  id::int, crm_id, nom, prenom, telephone, telephone_2, email, adresse, code_postal, ville, departement,
+  statut, telepro_email, telepro_name, notes, appels, dernier_appel_at, rdv_id::int,
   crm_statut, crm_campagne, crm_telepro, crm_commercial, crm_resultat_rdv, crm_commentaire, crm_source,
   crm_cree_le, crm_maj_le,
   to_char(dernier_rdv_date,'YYYY-MM-DD') as dernier_rdv_date, dernier_rdv_heure, dernier_rdv_presence,
